dedup load across persistences
