+++ src/Persistence/Sql.php
@@ -30,11 +30,6 @@ class Sql extends Persistence
     /** @const string */
     public const HOOK_BEFORE_DELETE_QUERY = self::class . '@beforeDeleteQuery';
 
-    /** @const string */
-    protected const ID_LOAD_ONE = self::class . '@idLoadOne';
-    /** @const string */
-    protected const ID_LOAD_ANY = self::class . '@idLoadAny';
-
     /**
      * Connection object.
      *
@@ -667,11 +662,6 @@ class Sql extends Persistence
         return $query;
     }
 
-    /**
-     * Tries to load data record, but will not fail if record can't be loaded.
-     *
-     * @param mixed $id
-     */
     public function tryLoad(Model $model, $id): ?array
     {
         $noId = $id === self::ID_LOAD_ONE || $id === self::ID_LOAD_ANY;
@@ -716,43 +706,6 @@ class Sql extends Persistence
         }
 
         return $data;
-    }
-
-    /**
-     * Loads a record from model and returns a associative array.
-     *
-     * @param mixed $id
-     */
-    public function load(Model $model, $id): array
-    {
-        $data = $this->tryLoad($model, $id);
-
-        if (!$data) {
-            $noId = $id === self::ID_LOAD_ONE || $id === self::ID_LOAD_ANY;
-
-            throw (new Exception('No record was found', 404))
-                ->addMoreInfo('model', $model)
-                ->addMoreInfo('id', $noId ? null : $id)
-                ->addMoreInfo('scope', $model->scope()->toWords());
-        }
-
-        return $data;
-    }
-
-    /**
-     * Tries to load any one record.
-     */
-    public function tryLoadAny(Model $model): ?array
-    {
-        return $this->tryLoad($model, self::ID_LOAD_ANY);
-    }
-
-    /**
-     * Loads any one record.
-     */
-    public function loadAny(Model $model): array
-    {
-        return $this->load($model, self::ID_LOAD_ANY);
     }
 
     /**
